backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==12.1.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
pandas
joblib
scikit-learn
numpy
pytesseract
boto3
PyMuPDF
easyocr
vaderSentiment
wordcloud
nltk
